JSI::JSON::Pointer is the new home of ::JSON::Schema::Pointer
