+++ lib/jsi/json.rb
@@ -3,5 +3,6 @@ module JSI
     autoload :Node, 'jsi/json/node'
     autoload :ArrayNode, 'jsi/json/node'
     autoload :HashNode, 'jsi/json/node'
+    autoload :Pointer, 'jsi/json/pointer'
   end
 end
