+++ lib/jsi/schema.rb
@@ -97,13 +97,13 @@ module JSI
         end
         if parent_auri.fragment
           # add onto the fragment
-          parent_id_path = ::JSON::Schema::Pointer.new(:fragment, '#' + parent_auri.fragment).reference_tokens
+          parent_id_path = JSI::JSON::Pointer.new(:fragment, '#' + parent_auri.fragment).reference_tokens
           path_from_id_node = parent_id_path + path_from_id_node
           parent_auri.fragment = nil
         #else: no fragment so parent_id good as is
         end
 
-        fragment = ::JSON::Schema::Pointer.new(:reference_tokens, path_from_id_node).fragment
+        fragment = JSI::JSON::Pointer.new(:reference_tokens, path_from_id_node).fragment
         schema_id = parent_auri.to_s + fragment
 
         schema_id
